FILE: extracts Dto types into separate files
