+++ libs/menu-matriarch/util/src/lib/dish-dto.ts
@@ -1,0 +1,6 @@
+import { Dish } from './dish.interface';
+
+export interface DishDto extends Omit<Dish, 'ingredients' | 'tags'> {
+  ingredientIds: string[];
+  tagIds: string[];
+}
