+++ libs/menu-matriarch/util/src/lib/meal-dto.ts
@@ -1,0 +1,6 @@
+import { Meal } from './meal.interface';
+
+export interface MealDto extends Omit<Meal, 'dishes' | 'tags'> {
+  dishIds: string[];
+  tagIds: string[];
+}
