+++ libs/menu-matriarch/util/src/lib/meal.interface.ts
@@ -9,8 +9,3 @@ export interface Meal {
   dishes: Dish[];
   tags: Tag[];
 }
-
-export interface MealDto extends Omit<Meal, 'dishes' | 'tags'> {
-  dishIds: string[];
-  tagIds: string[];
-}
