+++ libs/menu-matriarch/util/src/lib/menu-dto.ts
@@ -1,0 +1,3 @@
+import { Menu } from './menu.interface';
+
+export type MenuDto = Omit<Menu, 'entries' | 'orientation' | 'fallbackText'>;
