+++ libs/menu-matriarch/util/src/lib/tag-dto.ts
@@ -1,0 +1,3 @@
+import { Tag } from './tag.interface';
+
+export type TagDto = Tag;
